Refactor 3D results view into dynamic component
